requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# DocAssist – requirements.txt
# Install with:  pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────────────────────

# ── Web framework ──────────────────────────────────────────────────────────
streamlit>=1.32.0

# ── LLM / AI ───────────────────────────────────────────────────────────────
ollama                              # Local Ollama API client
langchain                           # LangChain for LLM integration

# ── Embeddings ─────────────────────────────────────────────────────────────
sentence-transformers>=2.6.0        # all-MiniLM-L6-v2 embedding model
torch>=2.0.0                        # Required by sentence-transformers

# ── Vector store ───────────────────────────────────────────────────────────
faiss-cpu>=1.7.4                    # FAISS similarity search (CPU build)

# ── Document parsing ───────────────────────────────────────────────────────
pypdf                               # PDF text extractor
python-docx>=1.1.0                  # Word document parser

# ── Utilities ──────────────────────────────────────────────────────────────
numpy>=1.24.0
requests>=2.25.0                    # HTTP requests for Ollama connectivity check
